suppressed downloading if file exists; added progress messages

+++ cerenkov3_data/edge_coexpedia.sh
@@ -1,7 +1,16 @@
 #!/bin/bash
 
-wget "http://www.coexpedia.org/dump/human_cx_net.zip"
-unzip -q human_cx_net.zip -d ./resource/Coexpedia/
-rm human_cx_net.zip
+DIRECTORY="./resource/Coexpedia"
+
+if [ -d "${DIRECTORY}/Hsa" ]; then
+    # Control will enter here if that folder exists.
+    echo "[Coexpedia] folder '${DIRECTORY}/Hsa' exists; no downloading"
+else
+    wget "http://www.coexpedia.org/dump/human_cx_net.zip"
+    unzip -q human_cx_net.zip -d ${DIRECTORY}
+    rm human_cx_net.zip
+fi
 
 python3 edge_coexpedia.py
+
+echo "[Coexpedia] done!"
